backend/requirements.txt
# FastAPI Core
fastapi==0.128.2
uvicorn[standard]==0.40.0

# Database
sqlalchemy[asyncio]==2.0.46
asyncpg==0.31.0
psycopg2-binary==2.9.11

# Validation & Settings
pydantic==2.12.5
pydantic-settings==2.12.0
dnspython==2.8.0
email-validator==2.3.0

# JWT & Security
python-jose[cryptography]==3.5.0
PyJWT[crypto]==2.11.0
cryptography==46.0.4

# HTTP Client
httpx==0.28.1

# Rate Limiting
slowapi==0.1.9

# Environment
python-dotenv==1.2.1
